Move `FeedImageDataStore` and `LocalFeedImageDataLoader` to production
